Fix randomness more.

+++ java/spykfunc/udfs/GaussRand.java
@@ -3,9 +3,9 @@ package spykfunc.udfs;
 import org.apache.spark.sql.api.java.UDF3;
 import org.apache.commons.math3.distribution.NormalDistribution;
 
-public class GaussRand extends Rand<NormalDistribution> implements UDF3<Short, Float, Float, Float> {
+public class GaussRand extends Rand<NormalDistribution> implements UDF3<Integer, Float, Float, Float> {
     @Override
-    public Float call(Short id, Float mean, Float sd) throws Exception {
+    public Float call(Integer id, Float mean, Float sd) throws Exception {
         return (float) (this.get(id, () -> new NormalDistribution(mean, sd)).sample());
     }
 }
